rbldns multi-base patch
http://qmail.jms1.net/djbdns/rbldns-patch.shtml
http://qmail.jms1.net/djbdns/djbdns-1.05-rbldns-multi-base.patch

Normally, rbldns uses the value of the BASE environment variable to tell which
domain the incoming queries are supposed to be within. If a query is received
for a name which is not in this domain, the query is ignored.

This patch extends rbldns so it will respond to queries for more than one
domain. For example, if you have a single set of IP addresses but you want it
to be visible as both "rbl.delete.net" and "rbl.dont-spam.us", this patch
makes it possible to run both blacklists under the same service, using the
same IP address. The service will respond to queries under either domain name.

  2008-01-18 I have sold the domain name "delete.net". If you are reading
  this, the "rbl.delete.net" zone no longer exists - if you wish to use the
  RBL, you will need to use the "rbl.dont-spam.us" name instead. In fact, it
  was the potential sale of this domain name which prompted me to write the
  patch to begin with.

Configuring the patch is easy - instead of setting BASE to a single domain
name, set BASE to a list of domain names, separated by the ":" character.
For example...

  cd /service/rbl.delete.net
  echo 'rbl.delete.net:rbl.dont-spam.us' > env/BASE
  svc -t .

This will change an existing rbldns service which answers queries for the
"rbl.delete.net" domain, so that it ALSO answers queries for the
"rbl.dont-spam.us" domain.

Note that there is only one list of IP blocks which belong to "the list"...
the change is that "the list" can now be queried using multiple names.

Signed-off-by: Giam Teck Choon <giamteckchoon@gmail.com>